requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
python-dotenv
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
binance-futures-connector
praw